kongtze-backend/requirements.txt
fastapi[standard]>=0.115.0
sqlalchemy>=2.0.35
alembic>=1.13.3
asyncpg>=0.30.0
pydantic>=2.9.2
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
google-generativeai>=0.8.3
pillow>=10.4.0
uvicorn[standard]>=0.32.0
